requirements.txt
discord~=1.0.1
pymongo~=3.11.2
git+https://github.com/rthalley/dnspython
python-dateutil~=2.8.1
python-dotenv~=0.19.0
discord-py-interactions~=3.0.2
certifi~=2021.5.30